Fix basic_test
